+++ kanban-board-core/src/main/java/me/ad/kanban/controller/StageController.java
@@ -1,5 +1,6 @@
 package me.ad.kanban.controller;
 
+import me.ad.kanban.auth.AuthorizationService;
 import me.ad.kanban.config.CustomMessageProperties;
 import me.ad.kanban.dto.TaskDto;
 import me.ad.kanban.dto.StageDto;
@@ -12,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,11 +26,13 @@ public class StageController {
     public static final Logger log = LoggerFactory.getLogger(TeamController.class);
     private final CustomMessageProperties message;
     private final StageService stageService;
+    private final AuthorizationService authorizationService;
 
     @Autowired
-    public StageController(CustomMessageProperties message, StageService StageService) {
+    public StageController(CustomMessageProperties message, StageService StageService, AuthorizationService authorizationService) {
         this.message = message;
         this.stageService = StageService;
+        this.authorizationService = authorizationService;
     }
 
     @GetMapping(path = {"", "/", "/all"}, produces = {MediaType.APPLICATION_JSON_VALUE})
@@ -56,8 +60,9 @@ public class StageController {
     @PostMapping(path = "/create", consumes = {MediaType.APPLICATION_JSON_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseStatus(HttpStatus.CREATED)
-    public StageDto createStage(@RequestBody StageDto StageDto) {
-        return stageService.saveOrUpdateStageByDto(StageDto);
+    public StageDto createStage(@RequestBody StageDto stageDto, Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfProject(auth, stageDto.getProjectId());
+        return stageService.saveOrUpdateStageByDto(stageDto);
     }
 
     @GetMapping(path = "/{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
@@ -72,17 +77,20 @@ public class StageController {
 
     @PutMapping(path = "/{id}/update", consumes = {MediaType.APPLICATION_JSON_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE})
-    public StageDto updateStage(@PathVariable("id") String id, @RequestBody StageDto StageDto) {
-        return stageService.updateStageById(id, StageDto);
+    public StageDto updateStage(@PathVariable("id") String id, @RequestBody StageDto stageDto, Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfStage(auth, id);
+        return stageService.updateStageById(id, stageDto);
     }
 
     @DeleteMapping(path = "/{id}")
-    public void deleteStageById(@PathVariable("id") String id) {
+    public void deleteStageById(@PathVariable("id") String id, Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfStage(auth, id);
         stageService.deleteStageById(id);
     }
 
-    @DeleteMapping(path = "/{id}", consumes = {MediaType.APPLICATION_JSON_VALUE})
-    public void deleteStageByIds(@RequestBody List<String> idList) {
+    @PostMapping(path = "/remove", consumes = {MediaType.APPLICATION_JSON_VALUE})
+    public void deleteStageByIdList(@RequestBody List<String> idList, Authentication auth) {
+        idList.forEach((id) -> authorizationService.ensureAuthUserIsOwnerOfStage(auth, id));
         stageService.deleteStages(idList);
     }
