+++ kanban-board-core/src/main/java/me/ad/kanban/controller/TeamController.java
@@ -1,5 +1,6 @@
 package me.ad.kanban.controller;
 
+import me.ad.kanban.auth.AuthorizationService;
 import me.ad.kanban.config.CustomMessageProperties;
 import me.ad.kanban.dto.TaskDto;
 import me.ad.kanban.dto.TeamDto;
@@ -11,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,11 +25,13 @@ public class TeamController {
     public static final Logger log = LoggerFactory.getLogger(TeamController.class);
     private final CustomMessageProperties message;
     private final TeamService teamService;
+    private final AuthorizationService authorizationService;
 
     @Autowired
-    public TeamController(CustomMessageProperties message, TeamService teamService) {
+    public TeamController(CustomMessageProperties message, TeamService teamService, AuthorizationService authorizationService) {
         this.message = message;
         this.teamService = teamService;
+        this.authorizationService = authorizationService;
     }
 
     @GetMapping(path = {"", "/", "/all"}, produces = {MediaType.APPLICATION_JSON_VALUE})
@@ -55,7 +59,8 @@ public class TeamController {
     @PostMapping(path = "/create", consumes = {MediaType.APPLICATION_JSON_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE})
     @ResponseStatus(HttpStatus.CREATED)
-    public TeamDto createTeam(@RequestBody TeamDto teamDto) {
+    public TeamDto createTeam(@RequestBody TeamDto teamDto, Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfProject(auth, teamDto.getProjectId());
         return teamService.saveOrUpdateTeamByDto(teamDto);
     }
 
@@ -76,27 +81,35 @@ public class TeamController {
 
     @PutMapping(path = "/{id}/update", consumes = {MediaType.APPLICATION_JSON_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE})
-    public TeamDto updateTeam(@PathVariable("id") String id, @RequestBody TeamDto teamDto) {
+    public TeamDto updateTeam(@PathVariable("id") String id, @RequestBody TeamDto teamDto,
+                              Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfTeam(auth, id);
         return teamService.updateTeamById(id, teamDto);
     }
 
     @DeleteMapping(path = "/{id}")
-    public void deleteTeamById(@PathVariable("id") String id) {
+    public void deleteTeamById(@PathVariable("id") String id, Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfTeam(auth, id);
         teamService.deleteTeamById(id);
     }
 
-    @DeleteMapping(path = "/{id}", consumes = {MediaType.APPLICATION_JSON_VALUE})
-    public void deleteTeamById(@RequestBody List<String> idList) {
+    @PostMapping(path = "/remove", consumes = {MediaType.APPLICATION_JSON_VALUE})
+    public void deleteTeamByIdList(@RequestBody List<String> idList, Authentication auth) {
+        idList.forEach((id) -> authorizationService.ensureAuthUserIsOwnerOfTeam(auth, id));
         teamService.deleteTeams(idList);
     }
 
     @PostMapping(path = "/{id}/users/add", consumes = {MediaType.APPLICATION_JSON_VALUE})
-    public void addUserToTeam(@PathVariable("id") String teamId, @RequestBody UserDto userDto) {
+    public void addUserToTeam(@PathVariable("id") String teamId, @RequestBody UserDto userDto,
+                              Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfTeam(auth, teamId);
         teamService.addUserToTeam(teamId, userDto);
     }
 
     @DeleteMapping(path = "/{id}/users/{userId}/remove")
-    public void removeUserFromTeam(@PathVariable("id") String teamId, @PathVariable("userId") String userId) {
+    public void removeUserFromTeam(@PathVariable("id") String teamId, @PathVariable("userId") String userId,
+                                   Authentication auth) {
+        authorizationService.ensureAuthUserIsOwnerOfTeam(auth, teamId);
         teamService.removeUserFromTeam(teamId, userId);
     }
 }
